move color correction to postprocess folder
