+++ src/ts/postProcesses/colorCorrection.ts
@@ -16,7 +16,7 @@
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import { PostProcess } from "@babylonjs/core/PostProcesses/postProcess";
-import colorCorrectionFragment from "../../../shaders/colorCorrection.glsl";
+import colorCorrectionFragment from "../../shaders/colorCorrection.glsl";
 import { Effect } from "@babylonjs/core/Materials/effect";
 import { Texture } from "@babylonjs/core/Materials/Textures/texture";
 import { Scene } from "@babylonjs/core/scene";
